Add some P_MobjWasRemoved checks around code that shrink touches
It's the only place in the vanilla game where K_DropItems
is called that isn't from a direct collision with the player
it's being called on. It's also a syncfail that doesn't happen
anywhere else, and I saw my sync state appear to get slightly
corrupted when it happened. Let's see if this fixes anything...