Custom download refusal message
This is a bonafide hack. Here's what's going on if you set a message:

All files are marked as downloadable.
When a file is requested that normally you would be able to download, it is
marked as a dummy file. This causes the file size to be set to one byte, so
that it finishes quickly. The node is then marked as having requested such a
file.
When the node is marked this way and requests to join after having downloaded
all the files, it is refused with the custom message.

Goddamn! That's a mouthful.