New system to prevent multiple fopen
Instead of FILE pointer stored in the transfer, they go in their own array. The number of nodes accessing a file is kept track of, and it's only fully closed after everyone who needs it is done with it.

Means that will unfortunately do a lot more seeking, but multiple users can download the same file now without it being implementation specific.